:recycle: move types into a shared directory.
