+++ src/feedsCollector/members.ts
@@ -1,4 +1,4 @@
-import type { Member } from "./types";
+import type { Member } from "@types";
 
 export const members: Member[] = [
   {
