--- src/feedsCollector/types.ts
@@ -1,31 +1,0 @@
-export type FeedItem = {
-  title: string;
-  link: string;
-  contentSnippet?: string;
-  isoDate?: string;
-  dateMiliSeconds: number;
-};
-
-export type Member = {
-  id: string;
-  name: string;
-  avatarSrc: string;
-  role?: string;
-  bio?: string;
-  sources?: string[];
-  includeUrlRegex?: string;
-  excludeUrlRegex?: string;
-  githubUsername?: string;
-  twitterUsername?: string;
-  websiteUrl?: string;
-};
-
-export type PostItem = {
-  authorId: string;
-  authorName: string;
-  title: string;
-  link: string;
-  contentSnippet?: string;
-  isoDate?: string;
-  dateMiliSeconds: number;
-};
