+++ src/types/feedItem.ts
@@ -1,0 +1,7 @@
+export type FeedItem = {
+  title: string;
+  link: string;
+  contentSnippet?: string;
+  isoDate?: string;
+  dateMiliSeconds: number;
+};
