+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export type { Member } from "./member";
+export type { FeedItem } from "./feedItem";
+export type { PostItem } from "./postItem";
