+++ src/types/member.ts
@@ -1,0 +1,13 @@
+export type Member = {
+  id: string;
+  name: string;
+  avatarSrc: string;
+  role?: string;
+  bio?: string;
+  sources?: string[];
+  includeUrlRegex?: string;
+  excludeUrlRegex?: string;
+  githubUsername?: string;
+  twitterUsername?: string;
+  websiteUrl?: string;
+};
