+++ src/types/postItem.ts
@@ -1,0 +1,9 @@
+export type PostItem = {
+  authorId: string;
+  authorName: string;
+  title: string;
+  link: string;
+  contentSnippet?: string;
+  isoDate?: string;
+  dateMiliSeconds: number;
+};
