+++ tsconfig.json
@@ -10,7 +10,8 @@
       "@pages/*": ["src/pages/*"],
       "@styles/*": ["src/styles/*"],
       "@config": ["src/config.ts"],
-      "@blogContents": [".contents"]
+      "@types": ["src/types"],
+      "@blogContents": [".contents"],
     }
   }
 }
